MQTTD.SubTree

Documentation

data SubTree a

MQTT Topic Subscription tree.

modify :: Monoid a => Filter -> (a -> a) -> SubTree a -> SubTree a

Modify a subscription for the given filter.

This will create structure along the path and will not clean up any unused structure.

add :: Monoid a => Filter -> a -> SubTree a -> SubTree a

Add a value at the given filter path.

find :: Monoid a => Topic -> SubTree a -> a

Find subscribers of a given topic.

findMap :: Monoid m => Topic -> (a -> m) -> SubTree a -> m

Find all matching subscribers

flatten :: SubTree a -> [(Filter, a)]

flatten a SubTree to a list of (topic,a) pairs. Note this will return mempty in paths along subscription paths.

fromList :: Monoid a => [(Filter, a)] -> SubTree a

Construct a SubTree from a list of filters and subscribers.