Scheduler

Documentation

type TimedQueue a = Map UTCTime [a]

add :: Ord a => UTCTime -> a -> TimedQueue a -> TimedQueue a

ready :: UTCTime -> TimedQueue a -> ([a], TimedQueue a)

next :: TimedQueue a -> Maybe UTCTime

newtype QueueRunner a

Constructors

QueueRunner

Fields

_tq :: TVar (TimedQueue a)

newRunner :: MonadIO m => m (QueueRunner a)

enqueue :: (HasStats m, Ord a, MonadIO m) => UTCTime -> a -> QueueRunner a -> m ()

run :: (HasStats m, MonadLogger m, MonadIO m) => (a -> m ()) -> QueueRunner a -> m ()

Run forever.

runOnce :: (HasStats m, MonadLogger m, MonadIO m) => (a -> m ()) -> QueueRunner a -> m ()

Block until an item might be ready and then run (and remove) all ready items. This will sometimes run 0 items. It shouldn't ever run any items that are scheduled for the future, and it shouldn't forget any items that are ready.

diffTimeToMicros :: NominalDiffTime -> Int